Build as a 64-bit shared library

+++ linuxdoom-1.10/v_video.h
@@ -49,7 +49,7 @@ extern	byte*		screens[5];
 extern  int	dirtybox[4];
 
 extern	byte	gammatable[5][256];
-extern	int	usegamma;
+extern	long	usegamma;
